modify: change cart-item into detail-product-section
